Deduplicate isPromise and unwrapFeature helpers
Move the internal isPromise and unwrapFeature helper functions to a
shared location in backend-plugin-api, and import them from
backend-app-api and backend-test-utils to eliminate drift risk.

Made-with: Cursor
Made-with: Cursor
