+++ .changeset/deduplicate-ispromise-unwrapfeature.md
@@ -1,0 +1,6 @@
+---
+'@backstage/backend-app-api': patch
+'@backstage/backend-test-utils': patch
+---
+
+Deduplicated internal `isPromise` and `unwrapFeature` helpers.
