+++ packages/backend-app-api/src/wiring/helpers.ts
@@ -14,26 +14,12 @@
  * limitations under the License.
  */
 
-import { BackendFeature } from '@backstage/backend-plugin-api';
-
-/** @internal */
-export function unwrapFeature(
-  feature: BackendFeature | { default: BackendFeature },
-): BackendFeature {
-  if ('$$type' in feature) {
-    return feature;
-  }
-
-  // This is a workaround where default exports get transpiled to `exports['default'] = ...`
-  // in CommonJS modules, which in turn results in a double `{ default: { default: ... } }` nesting
-  // when importing using a dynamic import.
-  // TODO: This is a broader issue than just this piece of code, and should move away from CommonJS.
-  if ('default' in feature) {
-    return feature.default;
-  }
-
-  return feature;
-}
+// Direct internal import to avoid duplication
+// eslint-disable-next-line @backstage/no-relative-monorepo-imports
+export {
+  isPromise,
+  unwrapFeature,
+} from '../../../backend-internal/src/wiring/helpers';
 
 /** @internal */
 export type DeepReadonly<T> = {
